refactor the code structure
