+++ code/fun-with-golang/intro/basic/basic.go
