+++ code/fun-with-golang/intro/hard/hard.go
