+++ code/fun-with-golang/intro/medium/medium.go
